Added Backround, Scrolling backround

+++ src/constants/gameConstants.py
@@ -5,5 +5,5 @@ ROWS= int(16)
 WINDOW_HEIGHT = int(WINDOW_WIDTH * 0.8)
 TILE_SIZE= float(WINDOW_HEIGHT / ROWS)
 TILE_TYPES = int(21)
-FPS = float(60)
+FPS = int(80)
